Migrate DI to Hilt
+ Some gradle and dependencies optimization and refactor
