+++ .gitignore
@@ -64,10 +64,6 @@ fabric.properties
 ### Intellij Patch ###
 # Comment Reason: https://github.com/joeblau/gitignore.io/issues/186#issuecomment-215987721
 
-# *.iml
-# modules.xml
-src/main/java/com/sadri/foursquare/ui/screens/dashboard/fragments/dashboard/explore/mvi/
-
 #.idea/.gitignore
 
 # ignore all .idea files ...
