+++ app/build.gradle
@@ -8,9 +8,9 @@ apply plugin: 'kotlin-kapt'
 
 apply plugin: "androidx.navigation.safeargs.kotlin"
 
-apply plugin: "de.mannodermaus.android-junit5"
-
 apply plugin: 'org.jlleitschuh.gradle.ktlint'
+
+apply plugin: 'dagger.hilt.android.plugin'
 
 android {
     compileSdkVersion rootProject.compileSdkVersion
@@ -25,6 +25,8 @@ android {
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         setProperty("archivesBaseName", "Four Square App v$versionName")
         vectorDrawables.useSupportLibrary true
+
+        testInstrumentationRunnerArguments clearPackageData: 'true'
     }
 
     buildTypes {
@@ -58,6 +60,7 @@ android {
     }
 
     testOptions {
+        execution 'ANDROIDX_TEST_ORCHESTRATOR'
         unitTests {
             includeAndroidResources = true
             returnDefaultValues = true
@@ -87,12 +90,11 @@ dependencies {
         exclude group: 'com.google.android.material', module: 'material'
     }
 
-    // Dagger
-    kapt "com.google.dagger:dagger-compiler:$rootProject.dagger"
-    kapt "com.google.dagger:dagger-android-processor:$rootProject.dagger"
-    implementation "com.google.dagger:dagger:$rootProject.dagger"
-    implementation "com.google.dagger:dagger-android:$rootProject.dagger"
-    implementation "com.google.dagger:dagger-android-support:$rootProject.dagger"
+    // Hilt
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    implementation "androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha03"
+    kapt "androidx.hilt:hilt-compiler:1.0.0-alpha03"
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
 
     // UI
     implementation "androidx.recyclerview:recyclerview:$rootProject.androidX.recyclerView"
@@ -114,16 +116,27 @@ dependencies {
     implementation "com.jakewharton.timber:timber:$rootProject.timber"
 
     // Test
-    testImplementation "junit:junit:$rootProject.testJUnit"
-    androidTestImplementation "androidx.test.ext:junit:$rootProject.testAndroidXJUnit"
-    androidTestImplementation "androidx.test.espresso:espresso-core:$rootProject.testEspresso"
-    testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:$rootProject.mockito"
-    testImplementation "org.mockito:mockito-inline:$rootProject.mockitoInline"
+    testImplementation "junit:junit:$rootProject.test.testJUnit"
+    testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:$rootProject.test.mockitoKotlin"
+    testImplementation "org.mockito:mockito-inline:$rootProject.test.mockitoInline"
     testImplementation "android.arch.core:core-testing:$rootProject.androidX.lifecycle"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$rootProject.coroutines"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$rootProject.coroutines"
-    kaptAndroidTest "com.google.dagger:dagger-compiler:$rootProject.dagger"
-    kaptTest "com.google.dagger:dagger-compiler:$rootProject.dagger"
+
+    // UI Test
+    androidTestImplementation "androidx.test:core:$rootProject.test.androidxCore"
+    androidTestImplementation "androidx.test.ext:junit:$rootProject.test.androidxJunit"
+    androidTestImplementation "androidx.test.ext:junit-ktx:$rootProject.test.androidxJunit"
+    androidTestImplementation "org.mockito:mockito-core:$rootProject.test.mockitoCore"
+
+    androidTestImplementation "androidx.test.espresso:espresso-core:$rootProject.test.espresso"
+    androidTestImplementation "androidx.test.espresso:espresso-intents:$rootProject.test.espresso"
+    androidTestImplementation "androidx.test.espresso:espresso-idling-resource:$rootProject.test.espresso"
+
+    androidTestImplementation "androidx.test:runner:$rootProject.test.runner"
+    androidTestImplementation "androidx.test.uiautomator:uiautomator:$rootProject.test.uiAutomator"
+    androidTestUtil "androidx.test:orchestrator:$rootProject.test.runner"
+
 
     //Gms Location
     implementation "com.google.android.gms:play-services-location:$rootProject.gmsLocation"
@@ -152,11 +165,5 @@ afterEvaluate {
             def flavorName = flavor.name.capitalize()
             tasks."compile${flavorName}DebugSources".dependsOn "ktlint${flavorName}DebugCheck"
             tasks."compile${flavorName}ReleaseSources".dependsOn "ktlint${flavorName}ReleaseCheck"
-    }
-}
-
-tasks.whenTaskAdded { task ->
-    if (task.name.contains("Test")) {
-        task.enabled = false
     }
 }
