--- app/src/main/java/com/sadri/foursquare/di/app/AppComponent.kt
@@ -1,39 +1,0 @@
-package com.sadri.foursquare.di.app
-
-import com.sadri.foursquare.App
-import com.sadri.foursquare.di.data.DataModule
-import com.sadri.foursquare.di.screen.ActivityModule
-import com.sadri.foursquare.di.screen.ViewModelsModule
-import dagger.BindsInstance
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-import javax.inject.Singleton
-
-/**
- * Created by Sepehr Sadri on 5/31/2020.
- * sepehrsadri@gmail.com
- * Tehran, Iran.
- * Copyright © 2020 by Sepehr Sadri. All rights reserved.
- */
-@Suppress("unused")
-@Singleton
-@Component(
-    modules = [
-        AndroidInjectionModule::class,
-        AppModule::class,
-        DataModule::class,
-        ActivityModule::class,
-        ViewModelsModule::class
-    ]
-)
-interface AppComponent {
-    @Component.Builder
-    interface Builder {
-        @BindsInstance
-        fun application(app: App): Builder
-
-        fun build(): AppComponent
-    }
-
-    fun inject(app: App)
-}
