--- app/src/main/java/com/sadri/foursquare/di/app/AppInjector.kt
@@ -1,18 +1,0 @@
-package com.sadri.foursquare.di.app
-
-import com.sadri.foursquare.App
-
-/**
- * Created by Sepehr Sadri on 5/31/2020.
- * sepehrsadri@gmail.com
- * Tehran, Iran.
- * Copyright © 2020 by Sepehr Sadri. All rights reserved.
- */
-object AppInjector {
-    fun init(app: App) {
-        DaggerAppComponent.builder()
-            .application(app)
-            .build()
-            .inject(app)
-    }
-}
