--- app/src/main/java/com/sadri/foursquare/di/app/Qualifier.kt
@@ -1,17 +1,0 @@
-package com.sadri.foursquare.di.app
-
-import javax.inject.Qualifier
-
-/**
- * Created by Sepehr Sadri on 5/31/2020.
- * sepehrsadri@gmail.com
- * Tehran, Iran.
- * Copyright © 2020 by Sepehr Sadri. All rights reserved.
- */
-@Suppress("unused")
-object Qualifier
-
-@Qualifier
-@MustBeDocumented
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ApplicationContext
