+++ app/src/main/java/com/sadri/foursquare/di/data/api/RetrofitModule.kt
@@ -3,8 +3,9 @@ package com.sadri.foursquare.di.data.api
 import com.sadri.foursquare.data.api.RetrofitProvider
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
-import javax.inject.Singleton
 
 /**
  * Created by Sepehr Sadri on 5/31/2020.
@@ -13,9 +14,9 @@ import javax.inject.Singleton
  * Copyright © 2020 by Sepehr Sadri. All rights reserved.
  */
 @Module
+@InstallIn(SingletonComponent::class)
 object RetrofitModule {
     @Provides
-    @Singleton
     fun provideRetrofit(): Retrofit {
         return RetrofitProvider.getRetrofit()
     }
