+++ app/src/main/java/com/sadri/foursquare/di/data/explore/VenueModule.kt
@@ -5,8 +5,9 @@ import com.sadri.foursquare.data.repositories.explore.ExploreServices
 import com.sadri.foursquare.data.repositories.venue_detail.VenueDetailServices
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
-import javax.inject.Singleton
 
 /**
  * Created by Sepehr Sadri on 6/3/2020.
@@ -15,9 +16,9 @@ import javax.inject.Singleton
  * Copyright © 2020 by Sepehr Sadri. All rights reserved.
  */
 @Module
+@InstallIn(SingletonComponent::class)
 object VenueModule {
     @Provides
-    @Singleton
     fun provideExploreApiServices(retrofit: Retrofit): ExploreServices {
         return RetrofitProvider.provideService(
             retrofit,
@@ -26,7 +27,6 @@ object VenueModule {
     }
 
     @Provides
-    @Singleton
     fun provideVenueDetailApiServices(retrofit: Retrofit): VenueDetailServices {
         return RetrofitProvider.provideService(
             retrofit,
