--- app/src/main/java/com/sadri/foursquare/di/utils/view_model/ViewModelFactory.kt
@@ -1,32 +1,0 @@
-package com.sadri.foursquare.di.utils.view_model
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import timber.log.Timber
-import javax.inject.Inject
-import javax.inject.Provider
-import javax.inject.Singleton
-
-/**
- * Created by Sepehr Sadri on 5/31/2020.
- * sepehrsadri@gmail.com
- * Tehran, Iran.
- * Copyright © 2020 by Sepehr Sadri. All rights reserved.
- */
-@Singleton
-class ViewModelFactory @Inject constructor(
-    private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val creator = creators[modelClass] ?: creators.entries.firstOrNull {
-            modelClass.isAssignableFrom(it.key)
-        }?.value ?: throw IllegalArgumentException("Unknown model class $modelClass")
-        try {
-            @Suppress("UNCHECKED_CAST")
-            return creator.get() as T
-        } catch (e: Exception) {
-            Timber.e(e, "Problem accrued with casting view model")
-            throw RuntimeException(e)
-        }
-    }
-}
