+++ app/src/main/java/com/sadri/foursquare/ui/screens/splash/SplashActivity.kt
@@ -2,8 +2,8 @@ package com.sadri.foursquare.ui.screens.splash
 
 import android.content.Intent
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.sadri.foursquare.ui.screens.dashboard.DashboardActivity
-import com.sadri.foursquare.ui.utils.BaseActivity
 
 /**
  * Created by Sepehr Sadri on 6/3/2020.
@@ -11,7 +11,7 @@ import com.sadri.foursquare.ui.utils.BaseActivity
  * Tehran, Iran.
  * Copyright © 2020 by Sepehr Sadri. All rights reserved.
  */
-class SplashActivity : BaseActivity() {
+class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
