--- app/src/main/java/com/sadri/foursquare/ui/utils/BaseActivity.kt
@@ -1,13 +1,0 @@
-package com.sadri.foursquare.ui.utils
-
-import android.annotation.SuppressLint
-import dagger.android.support.DaggerAppCompatActivity
-
-/**
- * Created by Sepehr Sadri on 5/31/2020.
- * sepehrsadri@gmail.com
- * Tehran, Iran.
- * Copyright © 2020 by Sepehr Sadri. All rights reserved.
- */
-@SuppressLint("Registered")
-abstract class BaseActivity : DaggerAppCompatActivity()
