+++ app/src/main/java/com/sadri/foursquare/ui/utils/DefaultDispatcher.kt
@@ -3,9 +3,7 @@ package com.sadri.foursquare.ui.utils
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
-import javax.inject.Singleton
 
-@Singleton
 class DefaultDispatcher @Inject constructor() : DispatcherProvider {
     override fun ui(): CoroutineDispatcher = Dispatchers.Main
     override fun io(): CoroutineDispatcher = Dispatchers.IO
