+++ app/src/main/java/com/sadri/foursquare/ui/utils/mvi/BaseMviActivity.kt
@@ -1,14 +1,14 @@
 package com.sadri.foursquare.ui.utils.mvi
 
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
-import com.sadri.foursquare.ui.utils.BaseActivity
 import com.sadri.foursquare.ui.utils.mvi.model.MviIntent
 import com.sadri.foursquare.ui.utils.mvi.model.MviViewState
 import timber.log.Timber
 
 abstract class BaseMviActivity<STATE : MviViewState, INTENT : MviIntent, ViewModel : BaseMviViewModel<STATE, INTENT, *>> :
-    BaseActivity() {
+    AppCompatActivity() {
 
     abstract val viewModel: ViewModel
