+++ build.gradle
@@ -2,15 +2,16 @@
 buildscript {
     ext.kotlin_version = "1.3.72"
     ext.navigation = '2.3.0-beta01'
+    ext.hilt_version = '2.32-alpha'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:4.0.0"
+        classpath 'com.android.tools.build:gradle:4.1.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$navigation"
-        classpath "de.mannodermaus.gradle.plugins:android-junit5:1.6.0.0"
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -44,7 +45,6 @@ ext {
     kotlin = '1.3.72'
 
     // App dependencies
-    dagger = '2.27'
     gson = '2.8.6'
     retrofit = '2.7.1'
     okHttp = '3.14.4'
@@ -53,15 +53,21 @@ ext {
     gmsLocation = '12.0.1'
     coroutines = '1.3.3'
 
-    // Test dependencies
-    testJUnit = '4.12'
-    testAndroidXJUnit = '1.1.1'
-    testEspresso = '3.2.0'
-    mockito = '2.2.0'
-    mockitoInline = '3.3.0'
-
     // ktLint
     ktlint = '9.2.1'
+
+    // Test
+    test = [
+            testJUnit    : '4.12',
+            espresso     : '3.1.0',
+            androidxJunit: '1.1.0',
+            androidxCore : "1.0.0",
+            mockitoInline: '3.3.0',
+            mockitoKotlin: '2.2.0',
+            mockitoCore  : "1.10.19",
+            runner       : '1.1.0',
+            uiAutomator  : '2.2.0',
+    ]
 
     // AndroidX
     androidX = [
